+++ test-suite/run_tests.sh
@@ -55,6 +55,7 @@ ALL_TEST_NAMES=(
     "TestBreakpointColumn"
     "TestGoto"
     "TestSingleThreadExec"
+    "TestEmbeddedSources"
 )
 
 TEST_NAMES="$@"